backend/requirements.txt
alembic==1.14.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.3
